Make examples friendlier for non GCC compilers and fixed slave_mem_i2c.c to compile on boards without I2C pins (#336)

+++ pwm/measure_duty_cycle/measure_duty_cycle.c
@@ -64,7 +64,7 @@ int main() {
     // values should be very close!
     for (int i = 0; i < count_of(test_duty_cycles); ++i) {
         float output_duty_cycle = test_duty_cycles[i];
-        pwm_set_gpio_level(OUTPUT_PIN, output_duty_cycle * (count_top + 1));
+        pwm_set_gpio_level(OUTPUT_PIN, (uint16_t) (output_duty_cycle * (count_top + 1)));
         float measured_duty_cycle = measure_duty_cycle(MEASURE_PIN);
         printf("Output duty cycle = %.1f%%, measured input duty cycle = %.1f%%\n",
                output_duty_cycle * 100.f, measured_duty_cycle * 100.f);
